[SV] More sentences to change brightness and colour of lights.
Added a few small adjustments for changing and setting brightness of lights and areas. And just a small change of colouring.

Basically just added "öka | minska" to previous "dimra" in the common file. Meaning to raise and lower the brightness. More fluent and compatible/bridgeable with/from previous smart assists.